+++ packages/moment/__tests__/moment.spec.ts
@@ -5,6 +5,16 @@ describe('bind moment', () => {
   it('parse correctly', () => {
     expect(parseToMoment('now-w/w+2d\\d').isSame(
       moment().subtract(1, 'week').startOf('week').add(2, 'days').endOf('day')
-    ));
+    )).toEqual(true);
+  });
+
+  it('compatible with grafana', () => {
+    console.log(parseToMoment('now-d/d', { forceEnd: true }));
+    expect(parseToMoment('now-d/d', { forceEnd: false }).isSame(
+      moment().subtract(1, 'd').startOf('d')
+    )).toEqual(true);
+    expect(parseToMoment('now-d/d', { forceEnd: true }).isSame(
+      moment().subtract(1, 'd').endOf('d')
+    )).toEqual(true);
   })
 });