+++ packages/moment/src/index.ts
@@ -4,8 +4,9 @@
 import moment from 'moment';
 import { parse } from 'relative-time-expression';
 
-export default function parseToMoment(exp: string) {
+export default function parseToMoment(exp: string, options?: { forceEnd?: boolean; }) {
   const ast = parse(exp);
+  const forceEnd = options && options.forceEnd;
   return ast.body.reduce((moment, m) => {
     if (m.type === 'Offset') {
       if (m.op === '+') {
@@ -14,7 +15,7 @@ export default function parseToMoment(exp: string) {
         return moment.subtract(m.number, m.unit);
       }
     } else {
-      if (m.op === '/') {
+      if (m.op === '/' && !forceEnd) {
         return moment.startOf(m.unit);
       } else {
         return moment.endOf(m.unit);