+++ HISTORY.md
@@ -1,2 +1,9 @@
+# 0.4.0
+- core: throw error object with infomation instead of message string.
+- moment-binding: add forceEnd to be compatible with grafana pattern.
+
+# 0.3.2 & 0.3.3
+- optimize error message
+
 # 0.3.1
 - bugfix: parse won't throw error `Cannot read property 'type' of undefined` when parse `"now-"`